+++ tools/bridge/src/tcp.rs
@@ -3,7 +3,7 @@ use std::net::{IpAddr}; // TODO: Add Ipv6Addr support
 use pnet::packet::tcp::{MutableTcpPacket,TcpFlags,TcpOption};
 
 
-use crate::protocol::db_data;
+use crate::drawbridge::db_data;
 
 
 // Builds an immutable TcpPacket to drop on the wire
